LP staking contract done

+++ contracts/PriceOracle/UniswapV2OracleLibrary.sol
@@ -10,7 +10,7 @@ library UniswapV2OracleLibrary {
 
     // helper function that returns the current block timestamp within the range of uint32, i.e. [0, 2**32 - 1]
     function currentBlockTimestamp() internal view returns (uint32) {
-        return uint32(block.timestamp % 2 ** 32);
+        return uint32(uint256(block.timestamp) % (2 ** 32));
     }
 
     // produces the cumulative price using counterfactuals to save gas and avoid a call to sync.
